[containers] change how missing cpu metrics are handled

+++ pkg/collector/corechecks/containers/docker/docker.go
@@ -344,10 +344,10 @@ func (d *DockerCheck) reportCPUMetrics(cpu *cmetrics.ContainerCPUStats, limits *
 		return
 	}
 
-	sender.Rate("docker.cpu.system", float64(cpu.System), "", tags)
-	sender.Rate("docker.cpu.user", float64(cpu.User), "", tags)
+	sender.Rate("docker.cpu.system", cpu.System, "", tags)
+	sender.Rate("docker.cpu.user", cpu.User, "", tags)
 	sender.Rate("docker.cpu.usage", cpu.UsageTotal, "", tags)
-	sender.Gauge("docker.cpu.shares", float64(cpu.Shares), "", tags)
+	sender.Gauge("docker.cpu.shares", cpu.Shares, "", tags)
 	sender.Rate("docker.cpu.throttled", float64(cpu.NrThrottled), "", tags)
 	sender.Rate("docker.cpu.throttled.time", cpu.ThrottledTime, "", tags)
 	if cpu.ThreadCount != 0 {
